Fix lint, move to adapters: PlayerStatsRecyclerViewAdapter
